feat(home): Added accordionPosts
