--- src/app/(home)/__components/__test__/Button.test.tsx
@@ -1,11 +1,0 @@
-import { fireEvent, render, screen } from "@testing-library/react";
-import Button from "../button";
-
-test("Calls onClick when button is clicked", () => {
-  const handleClick = jest.fn();
-
-  render(<Button onClick={handleClick}>Click Me!</Button>);
-  const button = screen.getByText("Click Me!");
-  fireEvent.click(button);
-  expect(handleClick).toHaveBeenCalledTimes(1);
-});
