+++ src/app/(home)/__components/accordionPosts.module.scss
@@ -1,0 +1,50 @@
+.accordionContainer {
+  width: 720px;
+  overflow: hidden;
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  .accordionItem {
+    list-style: none;
+    height: 500px;
+    transition: width 0.5s;
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center center;
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    &:not(:hover) {
+      width: 20%;
+    }
+    &:hover {
+      width: 600px;
+      &::after {
+        opacity: 1;
+      }
+    }
+    &::after {
+      content: "";
+      width: 100%;
+      height: 100%;
+      position: absolute;
+      left: 0;
+      top: 0;
+      background-color: rgba(0, 0, 0, 0.4);
+      z-index: 1;
+      opacity: 0;
+      transition: opacity 0.5s;
+    }
+    .accordionTitle {
+      font-size: 19px;
+      text-align: center;
+      color: white;
+      text-shadow: 2px 2px 4px #000000;
+      z-index: 2;
+      &:hover {
+        font-size: 21px;
+      }
+    }
+  }
+}
