--- src/app/(home)/__components/button.tsx
@@ -1,17 +1,0 @@
-import { FC, ReactNode } from "react";
-
-interface Props {
-  type?: "button" | "submit" | "reset";
-  onClick?: () => void;
-  className?: string;
-  children: ReactNode;
-}
-const Button: FC<Props> = ({ type, className, onClick, children }) => {
-  return (
-    <button type={type} onClick={onClick} className={className}>
-      {children}
-    </button>
-  );
-};
-
-export default Button;
